Performed refactoring
Moved profane_words to new file and added comments to files.
